requirements.txt
django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.1
django-ckeditor==6.7.0
django-storages==1.14.2
djangorestframework-simplejwt==5.3.1
django-anymail==10.3
gunicorn==21.2.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
Pillow==10.2.0
python-dotenv==1.0.1
